refact: Move all DI/container logic into dedicated class
As part of this effort, a basic Bootstrapper has been added, too.
