+++ composer.json
@@ -26,10 +26,12 @@
 	"autoload": {
 		"psr-4": {
 			"IndieHD\\AudioManipulator\\": "src/"
-		}
+		},
+		"files": ["src/Bootstrap.php"]
 	},
 	"minimum-stability": "stable",
 	"require-dev": {
-		"phpunit/phpunit": "^8.0"
+		"phpunit/phpunit": "^8.0",
+		"phpro/grumphp": "^0.15.0"
 	}
 }
