+++ src/Bootstrap.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace IndieHD\AudioManipulator;
+
+use IndieHD\AudioManipulator\Container;
+
+if (!defined('app')) {
+    function app()
+    {
+        return new Container();
+    }
+}
